:page_facing_up: Add a license file
Open-source projects often have a license file. This enables the creator of the project, in this case - Stellar, to tell people viewing the project what permissions they have over the source code of the project. I have chosen one of the more permissive licenses as it allows people to have more freedom!